Add DISABLE_SDIST_PROJECT_CHECK
By default fail if cabal.project/stack.yaml is found in the
distribution. Use this variable to ignore it and continue.

If we ignore the file without explicit intervention by the user then the
user might expect the build to use the packed project file, but we won't
use it. so ask the user permission so that they know.